pyproject.toml
[tool.poetry]
name = "cv-generator"
version = "2024"
description = "CV exports from yaml"
authors = ["Damien Allen <dev@dallen.co>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"

jinja2 = "^3.1.4"
pyyaml = "^6.0.1"
pydantic = "^2.6.4"
typer = "^0.12.3"


[tool.poetry.group.dev.dependencies]
ruff = "^0.4.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
